sequence for turning off body and time date check
The body was turning off because of where it was in the checking sequence. This is fixed now as I added all the checks just to the setup and added a boolean switch once checks were done. If no checks are true then in the setup it turns the display off.

Also in this iteration I have added time to the date check.